Add create-db script for sqlite and postgresql.

+++ bin/cmds/db.ts
@@ -1,32 +1,24 @@
 import { Database } from '@leafac/sqlite'
 import { execSync } from 'child_process'
 import fs from 'fs'
-import { createUser, prisma, sampleUsers } from '../../prisma/queries'
 
-export const createDatabase = async () => {
-  const dbPath = 'prisma/dev.db'
+export const createDb = async ({ type }: { type: 'postgresql' | 'sqlite' }) => {
+  if (type === 'sqlite') {
+    const dbPath = 'prisma/dev.db'
 
-  try {
-    // delete existing db
-    fs.unlinkSync(dbPath)
-  } catch (err) {
-    console.error('error removing database, does not exist.')
-  } finally {
-    // create db
-    new Database(dbPath)
-
-    // apply prisma schema
-    execSync('npx prisma migrate dev', { stdio: 'inherit' })
-
-    // load user data
-    for (let ii = 0; ii < sampleUsers.length; ii++) {
-      try {
-        console.log(await createUser(sampleUsers[ii]))
-      } catch (err) {
-        throw err
-      }
+    try {
+      // delete existing db
+      fs.unlinkSync(dbPath)
+    } catch (err) {
+      console.error('error removing database, does not exist.')
+    } finally {
+      new Database(dbPath)
     }
+  } else if (type === 'postgresql') {
   }
 
-  prisma.$disconnect()
+  // apply prisma schema
+  execSync('npx prisma migrate dev --name init-db', { stdio: 'inherit' })
 }
+
+export const seedDb = async () => {}
